Adds contract test for CREATE DATABASE for MySQL
